refact: Replaced Double atribute of balance with BigDecimal

+++ src/main/java/com/tiny/ledger/account/dto/AccountCreationRequest.java
@@ -1,7 +1,9 @@
 package com.tiny.ledger.account.dto;
 
+import java.math.BigDecimal;
+
 public record AccountCreationRequest(
          String name,
-         Double balance
+         BigDecimal balance
 ) {
 }
