+++ src/main/java/com/tiny/ledger/account/dto/AccountResponse.java
@@ -1,5 +1,7 @@
 package com.tiny.ledger.account.dto;
 
-public record AccountResponse(Long userId,String name, Double balance) {
+import java.math.BigDecimal;
+
+public record AccountResponse(Long userId, String name, BigDecimal balance) {
 
 }
