+++ src/main/java/com/tiny/ledger/transaction/dto/TransactionRequest.java
@@ -2,5 +2,7 @@ package com.tiny.ledger.transaction.dto;
 
 import com.tiny.ledger.transaction.domain.TransactionType;
 
-public record TransactionRequest(Double amount,TransactionType transactionType) {
+import java.math.BigDecimal;
+
+public record TransactionRequest(BigDecimal amount, TransactionType transactionType) {
 }
